requirements.txt
Django==2.1.5
djangorestframework==3.9.1
psycopg2==2.7.6.1
pytz==2018.9
whitenoise==4.1.2
